+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - post_check:  Added branch name as arg to is_commits_ahead and is_commits_behind calls.
 
 ### Changed
+- merge:  Added quarantine function call to quarantine new/modified files.
 - post_process:  Added datetime to destination directory for the archiving of the repo.
 - merge:  Added datetime to destination directory for the archiving of the repo.
 - merge_project:  Added additional logging error statements.