+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - post_check:  Added branch name as arg to is_commits_ahead and is_commits_behind calls.
 
 ### Changed
+- main:  Set the repo name for -r option if not passed to program using -p option setting.
 - merge:  Added quarantine function call to quarantine new/modified files.
 - prepare_mail:  Replaced post body mail lines with call to post_body function.
 - post_process:  Added datetime to destination directory for the archiving of the repo.